update layout, enhance clarity on thresholds
