+++ defi-alerts/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "compile": "prisma generate && tsc",
-    "dev": "prisma generate && next dev",
+    "dev": "prisma generate && next dev --turbopack",
     "build": "prisma generate && next build",
     "start": "prisma generate && next start",
     "lint": "next lint",
